requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==4.2.29
django-ckeditor==6.7.3
django-js-asset==2.2.0
idna==3.11
pillow==11.3.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
stripe==14.4.1
typing_extensions==4.15.0
urllib3==2.6.3
